# internal/controllers/dash_controller.go
package controllers

import (
	"github.com/HarshKanjiya/escape-form-api/internal/config"
	"github.com/HarshKanjiya/escape-form-api/internal/services"
	"github.com/gofiber/fiber/v2"
)

type DashController struct {
}

func NewDashController(*services.DashService, *config.Config) *DashController {
	return &DashController{}
}

// @Summary Get form analytics
// @Description Retrieve analytics for a form
// @Tags dashboard
// @Accept json
// @Produce json
// @Param formId path string true "Form ID"
// @Success 200 {object} map[string]interface{}
// @Router /dashboard/{formId}/analytics [get]
func (pc *DashController) GetAnalytics(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"message": "DashController Get method called",
	})
}

// @Summary Get form questions
// @Description Retrieve questions for a form
// @Tags dashboard
// @Accept json
// @Produce json
// @Param formId path string true "Form ID"
// @Success 200 {object} map[string]interface{}
// @Router /dashboard/{formId}/questions [get]
func (pc *DashController) GetQuestions(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"message": "DashController Get method called",
	})
}

// @Summary Get form responses
// @Description Retrieve responses for a form
// @Tags dashboard
// @Accept json
// @Produce json
// @Param formId path string true "Form ID"
// @Success 200 {object} map[string]interface{}
// @Router /dashboard/{formId}/responses [get]
func (pc *DashController) GetResponses(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"message": "DashController Get method called",
	})
}

// @Summary Get form passwords
// @Description Retrieve passwords for a form
// @Tags dashboard
// @Accept json
// @Produce json
// @Param formId path string true "Form ID"
// @Success 200 {object} map[string]interface{}
// @Router /dashboard/{formId}/passwords [get]
func (pc *DashController) GetPasswords(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"message": "DashController Get method called",
	})
}

// @Summary Update form passwords
// @Description Update passwords for a form
// @Tags dashboard
// @Accept json
// @Produce json
// @Param formId path string true "Form ID"
// @Param passwordId path string true "Password ID"
// @Success 200 {object} map[string]interface{}
// @Router /dashboard/{formId}/passwords/{passwordId} [patch]
func (pc *DashController) UpdatePasswords(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"message": "DashController UpdatePasswords method called",
	})
}

// @Summary Create form passwords
// @Description Create passwords for a form
// @Tags dashboard
// @Accept json
// @Produce json
// @Param formId path string true "Form ID"
// @Success 200 {object} map[string]interface{}
// @Router /dashboard/{formId}/passwords [post]
func (pc *DashController) CreatePasswords(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"message": "DashController UpdatePasswords method called",
	})
}

// @Summary Delete form passwords
// @Description Delete passwords for a form
// @Tags dashboard
// @Accept json
// @Produce json
// @Param formId path string true "Form ID"
// @Param passwordId path string true "Password ID"
// @Success 200 {object} map[string]interface{}
// @Router /dashboard/{formId}/passwords/{passwordId} [delete]
func (pc *DashController) DeletePasswords(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"message": "DashController DeletePasswords method called",
	})
}

func (pc *DashController) Create(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"message": "DashController Create method called",
	})
}

func (pc *DashController) Update(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"message": "DashController Update method called",
	})
}

func (pc *DashController) Delete(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"message": "DashController Delete method called",
	})
}

// @Summary Update form security
// @Description Update security settings for a form
// @Tags dashboard
// @Accept json
// @Produce json
// @Param formId path string true "Form ID"
// @Success 200 {object} map[string]interface{}
// @Router /dashboard/{formId}/security [patch]
func (pc *DashController) UpdateSecurity(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"message": "DashController UpdateSecurity method called",
	})
}

// @Summary Update form settings
// @Description Update settings for a form
// @Tags dashboard
// @Accept json
// @Produce json
// @Param formId path string true "Form ID"
// @Success 200 {object} map[string]interface{}
// @Router /dashboard/{formId}/settings [patch]
func (pc *DashController) UpdateSettings(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"message": "DashController UpdateSettings method called",
	})
}