# internal/controllers/edge_controller.go
package controllers

import (
	"github.com/HarshKanjiya/escape-form-api/internal/config"
	"github.com/HarshKanjiya/escape-form-api/internal/services"
	"github.com/gofiber/fiber/v2"
)

type EdgeController struct {
}

func NewEdgeController(*services.EdgeService, *config.Config) *EdgeController {
	return &EdgeController{}
}

// @Summary Get all edges
// @Description Retrieve a list of edges
// @Tags edges
// @Accept json
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Router /edges [get]
func (ec *EdgeController) Get(c *fiber.Ctx) error {
	// Placeholder for Get method implementation
	return c.JSON(fiber.Map{
		"message": "EdgeController Get method called",
	})
}

// @Summary Create a new edge
// @Description Create a new edge
// @Tags edges
// @Accept json
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Router /edges [post]
func (ec *EdgeController) Create(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"message": "EdgeController Create method called",
	})
}

// @Summary Update an edge
// @Description Update an existing edge by ID
// @Tags edges
// @Accept json
// @Produce json
// @Param id path string true "Edge ID"
// @Success 200 {object} map[string]interface{}
// @Router /edges/{id} [patch]
func (ec *EdgeController) Update(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"message": "EdgeController Update method called",
	})
}

// @Summary Delete an edge
// @Description Delete an edge by ID
// @Tags edges
// @Accept json
// @Produce json
// @Param id path string true "Edge ID"
// @Success 200 {object} map[string]interface{}
// @Router /edges/{id} [delete]
func (ec *EdgeController) Delete(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"message": "EdgeController Delete method called",
	})
}